refactor saving additional accounts
