--- template/accounts.json
@@ -1,52 +1,0 @@
-{
-    "account1":
-    {
-      "address": "cudos1yvtuaadhfhxf8ke7zm902z4rj382a8ayymq32s",
-      "mnemonic": "ordinary witness such toddler tag mouse helmet perfect venue eyebrow upgrade rabbit"
-    },
-    "account2":
-    {
-      "address": "cudos1xgejtl9heykzscska6cetsferpelygx8g9vxp9",
-      "mnemonic": "course hurdle stand heart rescue trap upset cousin dish embody business equip"
-    },
-    "account3":
-    {
-      "address": "cudos1u2f66a4g7yc95ndrjtjlh62qwxnrdxgntuxxkp",
-      "mnemonic": "warrior smoke melt opinion broccoli filter hole skate place cart ask alien"
-    },
-    "account4":
-    {
-      "address": "cudos1n8a8f580pjq4da82asd3tgdkwlxydecpaws8yg",
-      "mnemonic": "debate result toward brick dutch unusual caught plunge honey ramp want update"
-    },
-    "account5":
-    {
-      "address": "cudos1z44wsdcwrwdhjr57x7j65lultk4hm4ea9wfcaz",
-      "mnemonic": "tape soul absorb cabin luxury practice rally clerk love kiss throw avoid"
-    },
-    "account6":
-    {
-      "address": "cudos14f83hjnhxth3tmqtl7m7qycwwkzqlvzk708hum",
-      "mnemonic": "write senior detail business try dirt sort rural donate way acid flame"
-    },
-    "account7":
-    {
-      "address": "cudos15ew6kmnr0nk85agf0trglnk4kt37uk0xyl2qhl",
-      "mnemonic": "debris beef trip coral victory boy picture lab best royal wrestle path"
-    },
-    "account8":
-    {
-      "address": "cudos1hvrfp4r4l6aszvjx8la5lasef99kwvsf9d70lw",
-      "mnemonic": "damp bunker feed quit maze execute raccoon office error squirrel believe measure"
-    },
-    "account9":
-    {
-      "address": "cudos1lt9ky9e9h2npnrtk3mk9lqzekcc9q29q6t2lfv",
-      "mnemonic": "sea prison zero sugar episode meadow neutral silly chronic creek danger flat"
-    },
-    "account10":
-    {
-      "address": "cudos1j68t7ythea9fr9qt7ajf4nfc8hlw8p6420ll70",
-      "mnemonic": "save wear nature point april cage hybrid peasant penalty silver sadness friend"
-    }
-  }
